tidy: minor tidy. Mostly spelling, separate type per file.
